requirements.txt
absl-py==0.8.1
annoy==1.16.2
astor==0.8.0
blis==0.4.1
certifi==2019.9.11
chardet==3.0.4
cymem==2.0.2
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
more-itertools==7.2.0
murmurhash==1.0.2
numpy==1.17.4
plac==1.1.3
preshed==3.0.2
protobuf==3.10.0
requests==2.22.0
sentencepiece==0.1.83
simpleneighbors==0.0.1
six==1.13.0
spacy==2.2.2
srsly==0.2.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-hub==0.7.0
termcolor==1.1.0
tf-sentencepiece==0.1.83
thinc==7.3.1
tqdm==4.38.0
urllib3==1.25.7
wasabi==0.4.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
streamlit=>0.50.2